+++ CHANGELOG.rst
@@ -28,6 +28,7 @@ Fixed
 
 - Fixed defect where running ``dtool item properties`` with an invalid identifier
   resulted in a KeyError exception being propagated to the user
+- Fixed defect where ``dtool verify`` did not compare file sizes
 
 
 Security